Replace DashboardApp with DashboardAppNew
