+++ src/applications/personalization/dashboard/containers/DashboardApp.jsx
@@ -3,24 +3,27 @@ import { connect } from 'react-redux';
 import Scroll from 'react-scroll';
 import { withRouter } from 'react-router';
 
+import AlertBox from '@department-of-veterans-affairs/formation-react/AlertBox';
+
 import backendServices from 'platform/user/profile/constants/backendServices';
 import { selectProfile } from 'platform/user/selectors';
 import recordEvent from 'platform/monitoring/record-event';
 import localStorage from 'platform/utilities/storage/localStorage';
 
-import { removeSavedForm } from '../actions';
-
-import FormList from '../components/FormList';
-import MessagingWidget from './MessagingWidget';
-import ClaimsAppealsWidget from './ClaimsAppealsWidget';
-import PrescriptionsWidget from './PrescriptionsWidget';
-import PreferencesWidget from 'applications/personalization/preferences/containers/PreferencesWidget';
-
+import { removeSavedForm as removeSavedFormAction } from '../actions';
+import { getEnrollmentStatus as getEnrollmentStatusAction } from 'applications/hca/actions';
 import {
-  DowntimeNotification,
-  externalServices,
-} from 'platform/monitoring/DowntimeNotification';
-import AlertBox from '@department-of-veterans-affairs/formation-react/AlertBox';
+  hasServerError as hasESRServerError,
+  isEnrolledInVAHealthCare,
+} from 'applications/hca/selectors';
+
+import { recordDashboardClick } from '../helpers';
+
+import YourApplications from './YourApplications';
+import ManageYourVAHealthCare from '../components/ManageYourVAHealthCare';
+import ESRError, { ESR_ERROR_TYPES } from '../components/ESRError';
+import ClaimsAppealsWidget from './ClaimsAppealsWidget';
+import PreferencesWidget from 'applications/personalization/preferences/containers/PreferencesWidget';
 
 import profileManifest from 'applications/personalization/profile360/manifest.json';
 import accountManifest from 'applications/personalization/account/manifest.json';
@@ -36,32 +39,65 @@ const scrollToTop = () => {
   });
 };
 
-function recordDashboardClick(product, actionType = 'view-link') {
-  return () => {
-    recordEvent({
-      event: 'dashboard-navigation',
-      'dashboard-action': actionType,
-      'dashboard-product': product,
-    });
-  };
-}
-
-const ManageHealthBenefits = () => (
-  <>
-    <h2>Manage your health and benefits</h2>
+const EmptyStateLinks = () => (
+  <div>
+    <h2>Explore our most used benefits</h2>
     <ul className="va-nav-linkslist-list">
       <li>
         <a
-          href="/health-care/schedule-view-va-appointments/"
-          onClick={recordDashboardClick('schedule-appointment')}
+          href="/disability/"
+          onClick={recordDashboardClick('disability-benefits')}
         >
-          <h4 className="va-nav-linkslist-title">Schedule a VA appointment</h4>
+          <h4 className="va-nav-linkslist-title">Disability benefits</h4>
           <p className="va-nav-linkslist-description">
-            Find out how to make a doctor’s appointment with a member of your VA
-            health care team online or by phone.
+            Apply for disability compensation and other benefits for conditions
+            related to your military service.
           </p>
         </a>
       </li>
+      <li>
+        <a href="/health-care/" onClick={recordDashboardClick('health-care')}>
+          <h4 className="va-nav-linkslist-title">Health care benefits</h4>
+          <p className="va-nav-linkslist-description">
+            Apply for VA health care, find out how to access services, and
+            manage your health and benefits online.
+          </p>
+        </a>
+      </li>
+      <li>
+        <a
+          href="/education/"
+          onClick={recordDashboardClick('education-benefits')}
+        >
+          <h4 className="va-nav-linkslist-title">Education benefits</h4>
+          <p className="va-nav-linkslist-description">
+            Apply for and manage benefits that help you pay for college and
+            training programs.
+          </p>
+        </a>
+      </li>
+      <li>
+        <a
+          href="/careers-employment/"
+          onClick={recordDashboardClick('employment')}
+        >
+          <h4 className="va-nav-linkslist-title">Careers and employment</h4>
+          <p className="va-nav-linkslist-description">
+            Find out if you're eligible for Vocational Rehabilitation and
+            Employment (VR&E) services, get support for your Veteran-owned small
+            business, and access other resources to help build your career
+            skills and find a job.
+          </p>
+        </a>
+      </li>
+    </ul>
+  </div>
+);
+
+const ManageBenefitsOrRequestRecords = () => (
+  <>
+    <h2>Manage benefits or request records</h2>
+    <ul className="va-nav-linkslist-list">
       <li>
         <a
           href="/education/gi-bill/post-9-11/ch-33-benefit"
@@ -75,14 +111,6 @@ const ManageHealthBenefits = () => (
           </p>
         </a>
       </li>
-    </ul>
-  </>
-);
-
-const RequestYourRecords = () => (
-  <>
-    <h2>Request your records</h2>
-    <ul className="va-nav-linkslist-list">
       <li>
         <a
           href="/health-care/get-medical-records/"
@@ -161,6 +189,9 @@ class DashboardApp extends React.Component {
 
   componentDidMount() {
     scrollToTop();
+    if (this.props.profile.verified) {
+      this.props.getEnrollmentStatus();
+    }
   }
 
   dismissAlertBox = name => () => {
@@ -170,105 +201,7 @@ class DashboardApp extends React.Component {
     localStorage.setItem(`hide-${name}-alert`, true);
   };
 
-  renderWidgetDowntimeNotification = (appName, sectionTitle) => (
-    downtime,
-    children,
-  ) => {
-    switch (downtime.status) {
-      case 'down':
-        return (
-          <div>
-            <h2>{sectionTitle}</h2>
-            <AlertBox
-              content={
-                <div>
-                  <h4 className="usa-alert-heading">
-                    {appName} is down for maintenance
-                  </h4>
-                  <p>
-                    We’re making some updates to our {appName.toLowerCase()}{' '}
-                    tool. We’re sorry it’s not working right now and hope to be
-                    finished by {downtime.startTime.format('MMMM Do')},{' '}
-                    {downtime.endTime.format('LT')}. Please check back soon.
-                  </p>
-                </div>
-              }
-              isVisible
-              status="warning"
-            />
-          </div>
-        );
-      default:
-        return children;
-    }
-  };
-
-  renderEmptyStateLinks() {
-    return (
-      <div>
-        <h2>Explore our most used benefits</h2>
-
-        <ul className="va-nav-linkslist-list">
-          <li>
-            <a
-              href="/disability/"
-              onClick={recordDashboardClick('disability-benefits')}
-            >
-              <h4 className="va-nav-linkslist-title">Disability benefits</h4>
-              <p className="va-nav-linkslist-description">
-                Apply for disability compensation and other benefits for
-                conditions related to your military service.
-              </p>
-            </a>
-          </li>
-          <li>
-            <a
-              href="/health-care/"
-              onClick={recordDashboardClick('health-care')}
-            >
-              <h4 className="va-nav-linkslist-title">Health care benefits</h4>
-              <p className="va-nav-linkslist-description">
-                Apply for VA health care, find out how to access services, and
-                manage your health and benefits online.
-              </p>
-            </a>
-          </li>
-          <li>
-            <a
-              href="/education/"
-              onClick={recordDashboardClick('education-benefits')}
-            >
-              <h4 className="va-nav-linkslist-title">Education benefits</h4>
-              <p className="va-nav-linkslist-description">
-                Apply for and manage benefits that help you pay for college and
-                training programs.
-              </p>
-            </a>
-          </li>
-          <li>
-            <a
-              href="/careers-employment/"
-              onClick={recordDashboardClick('employment')}
-            >
-              <h4 className="va-nav-linkslist-title">Careers and employment</h4>
-              <p className="va-nav-linkslist-description">
-                Find out if you're eligible for Vocational Rehabilitation and
-                Employment (VR&E) services, get support for your Veteran-owned
-                small business, and access other resources to help build your
-                career skills and find a job.
-              </p>
-            </a>
-          </li>
-        </ul>
-      </div>
-    );
-  }
-
   renderLOAPrompt() {
-    if (this.props.profile.verified) {
-      return null;
-    }
-
     return (
       <AlertBox
         content={
@@ -288,7 +221,7 @@ class DashboardApp extends React.Component {
                 recordEvent({ event: 'verify-link-clicked' });
               }}
             >
-              Verify Your Identity
+              Verify your identity
             </a>
             <p>
               <a
@@ -310,13 +243,6 @@ class DashboardApp extends React.Component {
   }
 
   renderMVIWarning() {
-    if (
-      this.props.profile.loa.current === 1 ||
-      this.props.profile.status === 'OK'
-    ) {
-      return null;
-    }
-
     return (
       <AlertBox
         content={
@@ -340,16 +266,11 @@ class DashboardApp extends React.Component {
             <p>
               <a
                 href={facilityLocator.rootUrl}
-                onClick={() => {
-                  recordEvent({
-                    event: 'dashboard-navigation',
-                    'dashboard-action': 'view-link',
-                    'dashboard-product': 'find-center',
-                  });
-                }}
+                onClick={recordDashboardClick('find-center')}
               >
                 Find your nearest VA medical center
               </a>
+              .
             </p>
           </div>
         }
@@ -364,70 +285,63 @@ class DashboardApp extends React.Component {
   }
 
   render() {
+    const {
+      canAccessClaims,
+      canAccessRx,
+      canAccessMessaging,
+      canAccessAppeals,
+      profile,
+      showManageYourVAHealthCare,
+      showServerError,
+    } = this.props;
     const availableWidgetsCount = [
-      this.props.canAccessClaims,
-      this.props.canAccessRx,
-      this.props.canAccessMessaging,
-      this.props.canAccessAppeals,
+      canAccessClaims,
+      canAccessRx,
+      canAccessMessaging,
+      canAccessAppeals,
     ].filter(e => e).length;
 
     const view = (
-      <div className="row user-profile-row">
-        <div className="usa-width-two-thirds medium-8 small-12 columns">
-          <h1 id="dashboard-title">My VA</h1>
-          <div className="va-introtext">
-            <p>
-              Access the tools and information you’ll need to track and manage
-              your VA benefits and communications.
-            </p>
+      <>
+        <h1 id="dashboard-title">My VA</h1>
+        <div className="va-introtext">
+          <p>
+            Access the tools and information you’ll need to track and manage
+            your VA benefits and communications.
+          </p>
+        </div>
+
+        {showServerError && <ESRError errorType={ESR_ERROR_TYPES.generic} />}
+
+        <PreferencesWidget />
+
+        <YourApplications />
+
+        {!profile.verified && this.renderLOAPrompt()}
+        {profile.loa.current !== 1 &&
+          profile.status !== 'OK' &&
+          this.renderMVIWarning()}
+
+        <ClaimsAppealsWidget />
+
+        {availableWidgetsCount === 0 && <EmptyStateLinks />}
+
+        {showManageYourVAHealthCare && <ManageYourVAHealthCare />}
+        <ManageBenefitsOrRequestRecords />
+        <ViewYourProfile />
+        <ManageYourAccount />
+      </>
+    );
+
+    return (
+      <div name="topScrollElement">
+        <div className="row user-profile-row">
+          <div className="usa-width-two-thirds medium-8 small-12 columns">
+            {view}
           </div>
-
-          <PreferencesWidget />
-
-          <FormList
-            userProfile={this.props.profile}
-            removeSavedForm={this.props.removeSavedForm}
-            savedForms={this.props.profile.savedForms}
-          />
-
-          {this.renderLOAPrompt()}
-          {this.renderMVIWarning()}
-
-          <ClaimsAppealsWidget />
-
-          <DowntimeNotification
-            appTitle="messaging"
-            dependencies={[externalServices.mvi, externalServices.mhv]}
-            render={this.renderWidgetDowntimeNotification(
-              'Secure messaging',
-              'Track Secure Messages',
-            )}
-          >
-            <MessagingWidget />
-          </DowntimeNotification>
-
-          <DowntimeNotification
-            appTitle="rx"
-            dependencies={[externalServices.mvi, externalServices.mhv]}
-            render={this.renderWidgetDowntimeNotification(
-              'prescription refill',
-              'Refill Prescriptions',
-            )}
-          >
-            <PrescriptionsWidget />
-          </DowntimeNotification>
-
-          {availableWidgetsCount === 0 && this.renderEmptyStateLinks()}
-
-          <ManageHealthBenefits />
-          <RequestYourRecords />
-          <ViewYourProfile />
-          <ManageYourAccount />
         </div>
       </div>
     );
-
-    return <div name="topScrollElement">{view}</div>;
   }
 }
 
@@ -443,6 +357,7 @@ const mapStateToProps = state => {
   const canAccessClaims = profileState.services.includes(
     backendServices.EVSS_CLAIMS,
   );
+  const showServerError = hasESRServerError(state);
 
   return {
     canAccessRx,
@@ -450,11 +365,15 @@ const mapStateToProps = state => {
     canAccessAppeals,
     canAccessClaims,
     profile: profileState,
+    showManageYourVAHealthCare:
+      isEnrolledInVAHealthCare(state) || canAccessRx || canAccessMessaging,
+    showServerError,
   };
 };
 
 const mapDispatchToProps = {
-  removeSavedForm,
+  removeSavedForm: removeSavedFormAction,
+  getEnrollmentStatus: getEnrollmentStatusAction,
 };
 
 export default withRouter(
@@ -463,4 +382,4 @@ export default withRouter(
     mapDispatchToProps,
   )(DashboardApp),
 );
-export { DashboardApp };
+export { DashboardApp as DashboardAppNew };
