+++ src/applications/personalization/dashboard/routes.jsx
@@ -1,11 +1,6 @@
 import DashboardApp from './containers/DashboardApp';
-import DashboardAppNew from './containers/DashboardAppNew';
 import DashboardAppWrapper from './containers/DashboardAppWrapper';
 import SetPreferences from '../preferences/containers/SetPreferences';
-
-import environment from 'platform/utilities/environment';
-
-const component = environment.isProduction() ? DashboardApp : DashboardAppNew;
 
 export const findBenefitsRoute = {
   path: 'find-benefits',
@@ -17,7 +12,7 @@ export const findBenefitsRoute = {
 const routes = {
   path: '/',
   component: DashboardAppWrapper,
-  indexRoute: { component },
+  indexRoute: { component: DashboardApp },
   childRoutes: [findBenefitsRoute],
 };
