+++ src/applications/personalization/dashboard/tests/containers/DashboardApp.unit.spec.jsx
@@ -3,7 +3,7 @@ import SkinDeep from 'skin-deep';
 import sinon from 'sinon';
 import { expect } from 'chai';
 
-import localStorage from '../../../../../platform/utilities/storage/localStorage';
+import localStorage from 'platform/utilities/storage/localStorage';
 import { DashboardApp } from '../../containers/DashboardApp';
 
 const defaultProps = {
@@ -34,7 +34,7 @@ describe('<DashboardApp>', () => {
     const tree = SkinDeep.shallowRender(
       <DashboardApp profile={{ loa: { current: 1 }, verified: false }} />,
     );
-    expect(tree.toString()).to.contain('Explore our most used benefits');
+    expect(tree.toString()).to.contain('<EmptyStateLinks />');
   });
 
   it('should not render empty state links if there are available widgets', () => {
@@ -44,7 +44,7 @@ describe('<DashboardApp>', () => {
         profile={{ loa: { current: 1 }, verified: false }}
       />,
     );
-    expect(tree.toString()).not.to.contain('Explore Our Most Used Benefits');
+    expect(tree.toString()).not.to.contain('<EmptyStateLinks />');
   });
 
   it('should render verification state if LOA != 3', () => {
